NetworkClient и Api

+++ app/src/main/java/ru/practicum/android/diploma/commons/data/dto/search/VacanciesSearchByNameRequest.kt
@@ -5,5 +5,5 @@ import ru.practicum.android.diploma.commons.data.Constant.PER_PAGE_ITEMS
 data class VacanciesSearchByNameRequest(
     val name: String,
     val page: Long,
-    val amount: Int = PER_PAGE_ITEMS
+    val amount: Long = PER_PAGE_ITEMS
 )
